Implemented e2e tests for management class
Just implemented some e2e tests for the management class that cover
some of its main features. I also made some modification in the
management class.
